Add benchmarks, fix bad merge
