+++ doc/old/image_data_compression.rst
@@ -67,23 +67,22 @@ transfer syntaxes:
 .. _rle: https://github.com/pydicom/pylibjpeg-rle
 .. _gdcm: http://gdcm.sourceforge.net/
 
-+------------------------------------+--------------------+-------------------------+
-| Transfer Syntax                    | Plugin names       | Dependencies            |
-+--------------+---------------------+                    |                         |
-| Name         | UID                 |                    |                         |
-+==============+=====================+====================+=========================+
-| RLE Lossless | 1.2.840.10008.1.2.5 | pydicom            | `numpy <np_>`_          |
-+              +                     +--------------------+-------------------------+
-|              |                     | pylibjpeg :sup:`1` | `numpy <np_>`_,         |
-|              |                     |                    | `pylibjpeg <pylj_>`_,   |
-|              |                     |                    | `pylibjpeg-rle <rle_>`_ |
-+              +                     +--------------------+-------------------------+
-|              |                     | gdcm :sup:`2`      | `numpy <np_>`_,         |
-|              |                     |                    | `gdcm <gdcm_>`_,        |
-+--------------+---------------------+--------------------+-------------------------+
++------------------------------------+------------------+-------------------------+
+| Transfer Syntax                    | Plugin names     | Dependencies            |
++--------------+---------------------+                  |                         |
+| Name         | UID                 |                  |                         |
++==============+=====================+==================+=========================+
+| RLE Lossless | 1.2.840.10008.1.2.5 | pydicom :sup:`1` | `numpy <np_>`_          |
++              +                     +------------------+-------------------------+
+|              |                     | pylibjpeg        | `numpy <np_>`_,         |
+|              |                     |                  | `pylibjpeg <pylj_>`_,   |
+|              |                     |                  | `pylibjpeg-rle <rle_>`_ |
++              +                     +------------------+-------------------------+
+|              |                     | gdcm             | `numpy <np_>`_,         |
+|              |                     |                  | `gdcm <gdcm_>`_,        |
++--------------+---------------------+------------------+-------------------------+
 
-| :sup:`1` *~25x faster than the pydicom plugin*
-| :sup:`2` *~?x faster than the pydicom plugin*
+| :sup:`1` *~25x slower than the other plugins*
 
 Usage
 .....
