+++ doc/tutorials/pixel_data/compressing.rst
@@ -17,9 +17,18 @@ It's assumed that you're already familiar with the :doc:`dataset basics
 
 **Prerequisites**
 
+This tutorial uses packages in addition to *pydicom* that are not installed
+by default, but are required for *RLE Lossless* compression of *Pixel Data*.
+For more information on what packages are available to compress a given
+transfer syntax see the :doc:`image compression guide</old/image_data_compression>`.
+
+Installing using pip:
+
 .. code-block:: bash
 
     python -m pip install -U pydicom>=2.2 numpy pylibjpeg
+
+Installing on conda:
 
 .. code-block:: bash
