+++ pydicom/benchmarks/bench_rle_decode.py
