+++ pydicom/benchmarks/bench_rle_encode.py
@@ -7,6 +7,8 @@ from pydicom.pixel_data_handlers.rle_handler import (
     rle_encode_frame,
     _rle_encode_segment,
 )
+from pydicom.uid import RLELossless
+
 
 # 8/8-bit, 1 sample/pixel, 1 frame
 EXPL_8_1_1F = get_testdata_file("OBXXXX1A.dcm")
@@ -84,3 +86,35 @@ class TimeRLEEncodeFrame:
         """Time encoding 32 bit 3 sample/pixel."""
         for ii in range(self.no_runs):
             rle_encode_frame(self.arr32_3)
+
+
+# Requires numpy, pylibjpeg, pylibjpeg-rle and python-gdcm
+class TimeDatasetCompress:
+    """Test Dataset.compress()."""
+    def setup(self):
+        # More real-world like dataset
+        self.ds = dcmread(EXPL_8_1_1F)
+        self.arr8_1 = self.ds.pixel_array
+
+        self.no_runs = 1000
+
+    def time_pydicom(self):
+        """Time the native RLE encoder."""
+        for _ in range(self.no_runs):
+            self.ds.compress(
+                RLELossless, self.arr8_1, encoding_plugin='pydicom'
+            )
+
+    def time_pylibjpeg(self):
+        """Time the pylibjpeg-rle Rust RLE encoder."""
+        for _ in range(self.no_runs):
+            self.ds.compress(
+                RLELossless, self.arr8_1, encoding_plugin='pylibjpeg'
+            )
+
+    def time_gdcm(self):
+        """Time the GDCM C++ RLE encoder."""
+        for _ in range(self.no_runs):
+            self.ds.compress(
+                RLELossless, self.arr8_1, encoding_plugin='gdcm'
+            )
